pyproject.toml
[project]
name = "taxpose"
version = "0.1.0"
description = "TAX-Pose: Task-Specific Cross-Pose Estimation for Robot Manipulation"
readme = "README.md"
requires-python = ">=3.6, <=3.10.12" # ndf_robot requires < 3.10 :(
license = { file = "LICENSE.txt" }
authors = [
  { email = "chuerp@andrew.cmu.edu", name = "Chuer Pan" },
  { email = "bokorn@andrew.cmu.edu", name = "Brian Okorn" },
  { email = "haolunz@andrew.cmu.edu", name = "Harry Zhang" },
  { email = "baeisner@andrew.cmu.edu", name = "Ben Eisner" },
  { email = "dheld@andrew.cmu.edu", name = "David Held" },
]
dependencies = [
  # "chamferdist",
  "dgl ~= 1.0",
  "dglgo ~= 0.0.2",
  "h5py",
  "hydra-core",
  "imageio",
  "jupyterlab",
  "open3d",
  "opencv-python",
  "pybullet",
  "pyg_lib ~= 0.1; sys_platform == 'linux'",
  "pytorch3d ~= 0.7.2",
  "pytorch_lightning < 2.0",
  "rpad-partnet-mobility-utils @ git+https://github.com/r-pad/partnet_mobility_utils.git",
  "rpad-pyg @ git+https://github.com/r-pad/pyg_libs.git",
  "rpad-visualize-3d @ git+https://github.com/r-pad/visualize_3d.git",
  "wandb",
]

[build-system]
requires = ["setuptools >= 62.3.2", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
develop = [
  "autoflake",
  "black >= 22.3.0",
  "isort",
  "mypy >= 1.0",
  "pre-commit",
  "pylint",
  "pytest",
  "pytest-xdist",
  "types-tqdm",
]
rlbench = [
  "rpad-rlbench-utils @ git+https://github.com/r-pad/rlbench_utils.git",
]
build_docs = ["mkdocs-material", "mkdocstrings[python]"]


# This is required to allow us to have notebooks/ at the top level.
[tool.setuptools]
packages = ["taxpose", "third_party"]

[tool.setuptools.package-data]
taxpose = ["py.typed"]

[tool.black]
exclude = "third_party/*"

[tool.isort]
known_third_party = ["wandb"]
profile = "black"

[tool.mypy]
python_version = 3.9
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true

# Uncomment this when you have imports for mypy to ignore.
[[tool.mypy.overrides]]
module = [
  "chamferdist.*",
  "cv2.*",
  "display.*",
  "dgl.*",
  "h5py.*",
  "matplotlib.*",
  "pandas.*",
  "plotly.*",
  "pybullet.*",
  "pytorch3d.*",
  "scipy.*",
  "torch_cluster.*",
  "torch_geometric.*",
  "torchvision.*",
  "ompl.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--ignore=third_party/ -n auto"
testpaths = "tests"

[tool.pylint]
disable = [
  "C0103", # Invalid name.
  "C0114", # Missing module docstring.
  "C0115", # Missing class docstring.
  "C0116", # Missing function or method docstring.
]
known-third-party = ["wandb"]
